Added game state class and item names constants
Created a class for creating game state objects to pass between the backend and frontend of the application and eventually to a server for multi-player item tracking. In addition also defined all the names for each item as it will be referenced in the code base and a proper text value for pretty display if needed. Also went ahead and moved the endpoint callback anonymous function to its own proper place as I will be adding more logic and this will make it easier to test this function individually later.

Also added a custom rule to allow lines between class members. Having a blank space between each field just seems harder to read to me.
¯\_(ツ)_/¯