Update game state via Redux on emu POST request
I decided to use Redux to manage the state for the application. I've only added simple conditional logic for validating the request body was able to be parsed as JSON, but it's a decent proof of concept that can be expanded later on. Added a couple tests to ensure that the store is working correctly. Also there's a JS file that I use for debugging the lua script, but will be removed once I start work on the front-end part of the application as it's some hack-y stuff. Please don't read it.